updated the code from float to int

+++ RTMS_API/SensorCode.cs
@@ -15,7 +15,7 @@ namespace RTMS_API
     public partial class SensorCode
     {
         public int id { get; set; }
-        public Nullable<double> code { get; set; }
+        public Nullable<int> code { get; set; }
         public Nullable<System.DateTime> occurred { get; set; }
         public string notes { get; set; }
     }
